+++ .gitignore
@@ -14,3 +14,4 @@
 /captures
 .externalNativeBuild
 /app/google-services.json
+/alertdialog/google-services.json
