+++ alertdialog/src/main/java/com/crrl/beatplayer/alertdialog/stylers/AlertItemStyle.kt
@@ -11,10 +11,14 @@
  * limitations under the License.
  */
 
-package com.crrl.beatplayer.ui.widgets.stylers
+package com.crrl.beatplayer.alertdialog.stylers
 
-enum class AlertItemTheme {
-    DEFAULT,
-    CANCEL,
-    ACCEPT
-}
+import android.graphics.Color
+
+class AlertItemStyle(
+    var backgroundColor: Int = Color.parseColor("#F8F8F8"),
+    var selectedBackgroundColor: Int = Color.parseColor("#E8E8E8"),
+    var textColor: Int = Color.parseColor("#131313"),
+    var selectedTextColor: Int = Color.parseColor("#F44336"),
+    var cornerRadius: Float = 60f
+) : ItemStyle()
