+++ alertdialog/src/main/java/com/crrl/beatplayer/alertdialog/stylers/InputStyle.kt
@@ -11,10 +11,16 @@
  * limitations under the License.
  */
 
-package com.crrl.beatplayer.ui.widgets.stylers
+package com.crrl.beatplayer.alertdialog.stylers
 
-enum class AlertType {
-    BOTTOM_SHEET,
-    DIALOG,
-    INPUT
-}
+import android.graphics.Color
+
+class InputStyle(
+    var backgroundColor: Int = Color.parseColor("#131313"),
+    val inputColor: Int = Color.parseColor("#434343"),
+    val textColor: Int = Color.parseColor("#FFFFFF"),
+    val hintTextColor: Int = Color.parseColor("#AAAAAA"),
+    val acceptColor: Int = Color.parseColor("#00FF00"),
+    var text: String = "",
+    var cornerRadius: Float = 60f
+) : ItemStyle()
