+++ alertdialog/src/main/java/com/crrl/beatplayer/alertdialog/views/DialogAlert.kt
@@ -11,19 +11,19 @@
  * limitations under the License.
  */
 
-package com.crrl.beatplayer.ui.widgets.views
+package com.crrl.beatplayer.alertdialog.views
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.Button
-import com.crrl.beatplayer.R
-import com.crrl.beatplayer.interfaces.ItemListener
-import com.crrl.beatplayer.ui.widgets.actions.AlertItemAction
-import com.crrl.beatplayer.ui.widgets.stylers.AlertItemStyle
-import com.crrl.beatplayer.ui.widgets.stylers.AlertItemTheme
-import com.crrl.beatplayer.utils.GeneralUtils.drawRoundRectShape
+import com.crrl.beatplayer.alertdialog.R
+import com.crrl.beatplayer.alertdialog.interfaces.ItemListener
+import com.crrl.beatplayer.alertdialog.actions.AlertItemAction
+import com.crrl.beatplayer.alertdialog.stylers.AlertItemStyle
+import com.crrl.beatplayer.alertdialog.stylers.AlertItemTheme
+import com.crrl.beatplayer.alertdialog.utils.ViewUtils.drawRoundRectShape
 import com.google.android.material.bottomsheet.BottomSheetDialogFragment
 import kotlinx.android.synthetic.main.parent_dialog_layout.view.*
