+++ alertdialog/src/main/res/layout/dialog_item.xml
