+++ alertdialog/src/main/res/layout/input_dialog_item.xml
