+++ alertdialog/src/main/res/layout/parent_dialog_layout.xml
