+++ alertdialog/src/test/java/com/crrl/beatplayer/alertdialog/ExampleUnitTest.kt
@@ -11,6 +11,20 @@
  * limitations under the License.
  */
 
-package com.crrl.beatplayer.ui.widgets.stylers
+package com.crrl.beatplayer.alertdialog
 
-open class ItemStyle
+import org.junit.Test
+
+import org.junit.Assert.*
+
+/**
+ * Example local unit test, which will execute on the development machine (host).
+ *
+ * See [testing documentation](http://d.android.com/tools/testing).
+ */
+class ExampleUnitTest {
+    @Test
+    fun addition_isCorrect() {
+        assertEquals(4, 2 + 2)
+    }
+}
