+++ app/src/main/java/com/crrl/beatplayer/utils/GeneralUtils.kt
@@ -135,27 +135,6 @@ object GeneralUtils {
         }
     }
 
-    /**
-     * This method draws a round rect shape.
-     * @param width: int
-     * @param height: int
-     * @param color: int
-     * @return ShapeDrawable
-     */
-    fun drawRoundRectShape(
-        width: Int,
-        height: Int,
-        color: Int,
-        radius: Float = 30f
-    ): ShapeDrawable {
-        val r = floatArrayOf(radius, radius, radius, radius, radius, radius, radius, radius)
-        val oval = ShapeDrawable(RoundRectShape(r, RectF(), r))
-        oval.intrinsicHeight = height
-        oval.intrinsicWidth = width
-        oval.paint.color = color
-        return oval
-    }
-
     fun getAlbumArtBitmap(context: Context, albumId: Long?): Bitmap? {
         if (albumId == null) return null
         return try {
