+++ settings.gradle
@@ -10,5 +10,6 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+include ':alertdialog'
 
 include ':app'
